Added Unit Controller

+++ KejaHUnt_PropertiesAPI/Models/Domain/Unit.cs
@@ -4,7 +4,7 @@ namespace KejaHUnt_PropertiesAPI.Models.Domain
 {
     public class Unit
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         [Precision(18, 2)]
         public decimal Price { get; set; }
         public string Type { get; set; }
@@ -12,7 +12,7 @@ namespace KejaHUnt_PropertiesAPI.Models.Domain
         public double Size { get; set; }
         public int NoOfUnits { get; set; }
         public Guid? DocumentId { get; set; }
-        public int PropertyId { get; set; }
+        public long PropertyId { get; set; }
         public Property Property { get; set; }
 
     }
